requirements.txt
flask
PyPDF2
pillow
fpdf
lxml
